Add close failover version to history blob tags and support indexing of these versions
